A little bit of refactoring
