--- my_src/CMakeLists.txt
@@ -1,12 +1,0 @@
-set(SOURCE
-    ${SOURCE}
-    PARENT_SCOPE
-)
-
-set(HEADERS
-    ${HEADERS}
-    ${CMAKE_CURRENT_SOURCE_DIR}/
-    PARENT_SCOPE
-)
-
-message(STATUS "child :" ${SOURCE} ${HEADERS} ${TEST_SOURCES})
